Refactored file structure to split generic components into components/ and page/use specific components into features/
