+++ src/pages/index.tsx
@@ -1,9 +1,9 @@
-import { queryParamToFilterState } from "components/CourtCaseFilters/ResultFilter"
-import CourtCaseFilter from "components/CourtCaseFilter"
+import { queryParamToFilterState } from "features/CourtCaseFilters/ResultFilter"
+import CourtCaseFilter from "features/CourtCaseFilters/CourtCaseFilter"
 import AuthenticationServerSidePropsContext from "types/AuthenticationServerSidePropsContext"
 import { Filter, QueryOrder } from "types/CaseListQueryParams"
 import { isError } from "types/Result"
-import CourtCaseList from "components/CourtCaseList/CourtCaseList"
+import CourtCaseList from "features/CourtCaseList/CourtCaseList"
 import Layout from "components/Layout"
 import Pagination from "components/Pagination/Pagination"
 import User from "entities/User"
@@ -12,7 +12,7 @@ import { withAuthentication, withMultipleServerSideProps } from "middleware"
 import type { GetServerSidePropsContext, GetServerSidePropsResult, NextPage } from "next"
 import Head from "next/head"
 import { ParsedUrlQuery } from "querystring"
-import listCourtCases from "useCases/listCourtCases"
+import listCourtCases from "services/listCourtCases"
 
 import type CourtCase from "entities/CourtCase"
